Moved code from Webservice to Core
