+++ src/FunderMaps.Core/Interfaces/IUserTrackingService.cs
@@ -1,9 +1,10 @@
-﻿namespace FunderMaps.Webservice.Abstractions.Services
+﻿namespace FunderMaps.Core.Interfaces
 {
     /// <summary>
     /// Contract for tracking user behaviour regarding requests.
     /// </summary>
     public interface IUserTrackingService
     {
+        // TODO Implement
     }
 }
