+++ src/FunderMaps.Core/Types/ExternalDataSource.cs
@@ -1,0 +1,18 @@
+﻿namespace FunderMaps.Core.Types
+{
+    /// <summary>
+    /// Represents an external data source.
+    /// </summary>
+    public enum ExternalDataSource
+    {
+        /// <summary>
+        /// Basis Registratie Gebouwen.
+        /// </summary>
+        Bag,
+
+        /// <summary>
+        /// Open Street Maps.
+        /// </summary>
+        Osm,
+    }
+}
