+++ src/FunderMaps.Core/Types/FoundationRisk.cs
@@ -1,0 +1,33 @@
+﻿namespace FunderMaps.Core.Types
+{
+    /// <summary>
+    /// Represents the foundation risk.
+    /// </summary>
+    public enum FoundationRisk
+    {
+        /// <summary>
+        /// Risk level A.
+        /// </summary>
+        A,
+
+        /// <summary>
+        /// Risk level B.
+        /// </summary>
+        B,
+
+        /// <summary>
+        /// Risk level C.
+        /// </summary>
+        C,
+
+        /// <summary>
+        /// Risk level D.
+        /// </summary>
+        D,
+
+        /// <summary>
+        /// Risk level E.
+        /// </summary>
+        E
+    }
+}
