+++ src/FunderMaps.Core/Types/Years.cs
@@ -1,4 +1,4 @@
-﻿namespace FunderMaps.Webservice.Types
+﻿namespace FunderMaps.Core.Types
 {
     /// <summary>
     /// Represents a collection of years.
